+++ packages/contracts/test/registry/aragon-plugin-registry.ts
@@ -3,6 +3,8 @@ import {ethers} from 'hardhat';
 
 import {DAO, PluginRepo} from '../../typechain';
 import {customError} from '../test-utils/custom-error-helper';
+import {deployNewDAO} from '../test-utils/dao';
+import {deployNewPluginRepo} from '../test-utils/repo';
 
 const EVENTS = {
   PluginRepoRegistered: 'PluginRepoRegistered',
@@ -11,8 +13,11 @@ const EVENTS = {
 describe('Aragon-Plugin-Registry', function () {
   let aragonPluginRegistry: any;
   let ownerAddress: string;
-  let dao: DAO;
+  let managingDAO: DAO;
   let pluginRepo: PluginRepo;
+
+  const REGISTER_PERMISSION_ID = ethers.utils.id('REGISTER_PERMISSION_ID');
+  const pluginRepoName = 'my-pluginRepo';
 
   before(async () => {
     const signers = await ethers.getSigners();
@@ -21,33 +26,27 @@ describe('Aragon-Plugin-Registry', function () {
 
   beforeEach(async function () {
     // DAO
-    const DAO = await ethers.getContractFactory('DAO');
-    dao = await DAO.deploy();
-    await dao.initialize('0x00', ownerAddress, ethers.constants.AddressZero);
+    managingDAO = await deployNewDAO(ownerAddress);
 
     // deploy and initialize AragonPluginRegistry
     const AragonPluginRegistry = await ethers.getContractFactory(
       'AragonPluginRegistry'
     );
     aragonPluginRegistry = await AragonPluginRegistry.deploy();
-    await aragonPluginRegistry.initialize(dao.address);
+    await aragonPluginRegistry.initialize(managingDAO.address);
 
     // deploy a pluginRepo and initialize
-    const PluginRepo = await ethers.getContractFactory('PluginRepo');
-    pluginRepo = await PluginRepo.deploy();
-    await pluginRepo.initialize(ownerAddress);
+    pluginRepo = await deployNewPluginRepo(ownerAddress);
 
     // grant REGISTER_PERMISSION_ID to registrer
-    dao.grant(
+    managingDAO.grant(
       aragonPluginRegistry.address,
       ownerAddress,
-      ethers.utils.id('REGISTER_PERMISSION_ID')
+      REGISTER_PERMISSION_ID
     );
   });
 
   it('Should register a new pluginRepo successfully', async function () {
-    const pluginRepoName = 'my-pluginRepo';
-
     await expect(
       await aragonPluginRegistry.register(pluginRepoName, pluginRepo.address)
     )
@@ -59,9 +58,35 @@ describe('Aragon-Plugin-Registry', function () {
     );
   });
 
-  it('Should revert if pluginRepo already exists', async function () {
-    const pluginRepoName = 'my-pluginRepo';
+  it('fail to register if the sender lacks the required role', async () => {
+    // Register a plugin successfully
+    await aragonPluginRegistry.register(pluginRepoName, pluginRepo.address);
 
+    // Revoke the permission
+    await managingDAO.revoke(
+      aragonPluginRegistry.address,
+      ownerAddress,
+      REGISTER_PERMISSION_ID
+    );
+
+    // deploy a pluginRepo
+    const newPluginRepo = await deployNewPluginRepo(ownerAddress);
+
+    await expect(
+      aragonPluginRegistry.register(pluginRepoName, newPluginRepo.address)
+    ).to.be.revertedWith(
+      customError(
+        'DAOPermissionMissing',
+        managingDAO.address,
+        aragonPluginRegistry.address,
+        aragonPluginRegistry.address,
+        ownerAddress,
+        REGISTER_PERMISSION_ID
+      )
+    );
+  });
+
+  it('fail to register if pluginRepo already exists', async function () {
     await aragonPluginRegistry.register(pluginRepoName, pluginRepo.address);
 
     await expect(