+++ packages/subgraph/tests/registry/utils.ts
@@ -1,17 +1,16 @@
 import {Address, ethereum} from '@graphprotocol/graph-ts';
 import {newMockEvent} from 'matchstick-as/assembly/index';
 
-import {NewDAORegistered} from '../../generated/Registry/Registry';
+import {DAORegistered} from '../../generated/Registry/Registry';
 
 // events
 
 export function createNewDaoEvent(
   dao: string,
   creator: string,
-  token: string,
   name: string
-): NewDAORegistered {
-  let newDaoEvent = changetype<NewDAORegistered>(newMockEvent());
+): DAORegistered {
+  let newDaoEvent = changetype<DAORegistered>(newMockEvent());
 
   newDaoEvent.parameters = [];
 
@@ -23,10 +22,6 @@ export function createNewDaoEvent(
     'creator',
     ethereum.Value.fromAddress(Address.fromString(creator))
   );
-  let tokenParam = new ethereum.EventParam(
-    'token',
-    ethereum.Value.fromAddress(Address.fromString(token))
-  );
   let nameParam = new ethereum.EventParam(
     'name',
     ethereum.Value.fromString(name)
@@ -34,7 +29,6 @@ export function createNewDaoEvent(
 
   newDaoEvent.parameters.push(daoParam);
   newDaoEvent.parameters.push(creatorParam);
-  newDaoEvent.parameters.push(tokenParam);
   newDaoEvent.parameters.push(nameParam);
 
   return newDaoEvent;